output_server/requirements.txt
uvicorn==0.30.0
fastapi==0.103.0
psycopg2-binary==2.9.9